Retrieve project API key if stored, if not, set and retrieve

Attempting to make secure API key storage so simple, that no other way makes sense. Wrapping key_get and key_set using the key_list to check if key is in storage already.

Usage

get_api_key(key.name, ...)

Arguments

key.name

character vector of key name

...

passed to key_set

Value

character vector