Data set to data dictionary function

Migrated from stRoke ds2dd(). Fits better with the functionality of 'REDCapCAST'

Usage

ds2dd(
  ds,
  record.id = "record_id",
  form.name = "basis",
  field.type = "text",
  field.label = NULL,
  include.column.names = FALSE,
  metadata = names(redcapcast_meta)
)

Arguments

ds

data set

record.id

name or column number of id variable, moved to first row of data dictionary, character of integer. Default is "record_id".

form.name

vector of form names, character string, length 1 or length equal to number of variables. Default is "basis".

field.type

vector of field types, character string, length 1 or length equal to number of variables. Default is "text.

field.label

vector of form names, character string, length 1 or length equal to number of variables. Default is NULL and is then identical to field names.

include.column.names

Flag to give detailed output including new column names for original data set for upload.

metadata

Metadata column names. Default is the included REDCapCAST::redcapcast_data.

Value

data.frame or list of data.frame and vector

Examples

redcapcast_data$record_id <- seq_len(nrow(redcapcast_data))
ds2dd(redcapcast_data, include.column.names=TRUE)
#> $DataDictionary
#>                      field_name form_name section_header field_type
#> 1                     record_id     basis             NA       text
#> 2             redcap_event_name     basis             NA       text
#> 3      redcap_repeat_instrument     basis             NA       text
#> 4        redcap_repeat_instance     basis             NA       text
#> 5                           cpr     basis             NA       text
#> 6                     inclusion     basis             NA       text
#> 7                           dob     basis             NA       text
#> 8                           age     basis             NA       text
#> 9                   age_integer     basis             NA       text
#> 10                          sex     basis             NA       text
#> 11                 cohabitation     basis             NA       text
#> 12                 hypertension     basis             NA       text
#> 13                     diabetes     basis             NA       text
#> 14                       region     basis             NA       text
#> 15 baseline_data_start_complete     basis             NA       text
#> 16                 mrs_assessed     basis             NA       text
#> 17                     mrs_date     basis             NA       text
#> 18                    mrs_score     basis             NA       text
#> 19                 mrs_complete     basis             NA       text
#> 20                   event_date     basis             NA       text
#> 21                   event_type     basis             NA       text
#> 22           new_event_complete     basis             NA       text
#>                     field_label select_choices_or_calculations field_note
#> 1                     record_id                             NA         NA
#> 2             redcap_event_name                             NA         NA
#> 3      redcap_repeat_instrument                             NA         NA
#> 4        redcap_repeat_instance                             NA         NA
#> 5                           cpr                             NA         NA
#> 6                     inclusion                             NA         NA
#> 7                           dob                             NA         NA
#> 8                           age                             NA         NA
#> 9                   age_integer                             NA         NA
#> 10                          sex                             NA         NA
#> 11                 cohabitation                             NA         NA
#> 12                 hypertension                             NA         NA
#> 13                     diabetes                             NA         NA
#> 14                       region                             NA         NA
#> 15 baseline_data_start_complete                             NA         NA
#> 16                 mrs_assessed                             NA         NA
#> 17                     mrs_date                             NA         NA
#> 18                    mrs_score                             NA         NA
#> 19                 mrs_complete                             NA         NA
#> 20                   event_date                             NA         NA
#> 21                   event_type                             NA         NA
#> 22           new_event_complete                             NA         NA
#>    text_validation_type_or_show_slider_number text_validation_min
#> 1                                          NA                  NA
#> 2                                          NA                  NA
#> 3                                          NA                  NA
#> 4                                          NA                  NA
#> 5                                          NA                  NA
#> 6                                          NA                  NA
#> 7                                          NA                  NA
#> 8                                          NA                  NA
#> 9                                          NA                  NA
#> 10                                         NA                  NA
#> 11                                         NA                  NA
#> 12                                         NA                  NA
#> 13                                         NA                  NA
#> 14                                         NA                  NA
#> 15                                         NA                  NA
#> 16                                         NA                  NA
#> 17                                         NA                  NA
#> 18                                         NA                  NA
#> 19                                         NA                  NA
#> 20                                         NA                  NA
#> 21                                         NA                  NA
#> 22                                         NA                  NA
#>    text_validation_max identifier branching_logic required_field
#> 1                   NA         NA              NA             NA
#> 2                   NA         NA              NA             NA
#> 3                   NA         NA              NA             NA
#> 4                   NA         NA              NA             NA
#> 5                   NA         NA              NA             NA
#> 6                   NA         NA              NA             NA
#> 7                   NA         NA              NA             NA
#> 8                   NA         NA              NA             NA
#> 9                   NA         NA              NA             NA
#> 10                  NA         NA              NA             NA
#> 11                  NA         NA              NA             NA
#> 12                  NA         NA              NA             NA
#> 13                  NA         NA              NA             NA
#> 14                  NA         NA              NA             NA
#> 15                  NA         NA              NA             NA
#> 16                  NA         NA              NA             NA
#> 17                  NA         NA              NA             NA
#> 18                  NA         NA              NA             NA
#> 19                  NA         NA              NA             NA
#> 20                  NA         NA              NA             NA
#> 21                  NA         NA              NA             NA
#> 22                  NA         NA              NA             NA
#>    custom_alignment question_number matrix_group_name matrix_ranking
#> 1                NA              NA                NA             NA
#> 2                NA              NA                NA             NA
#> 3                NA              NA                NA             NA
#> 4                NA              NA                NA             NA
#> 5                NA              NA                NA             NA
#> 6                NA              NA                NA             NA
#> 7                NA              NA                NA             NA
#> 8                NA              NA                NA             NA
#> 9                NA              NA                NA             NA
#> 10               NA              NA                NA             NA
#> 11               NA              NA                NA             NA
#> 12               NA              NA                NA             NA
#> 13               NA              NA                NA             NA
#> 14               NA              NA                NA             NA
#> 15               NA              NA                NA             NA
#> 16               NA              NA                NA             NA
#> 17               NA              NA                NA             NA
#> 18               NA              NA                NA             NA
#> 19               NA              NA                NA             NA
#> 20               NA              NA                NA             NA
#> 21               NA              NA                NA             NA
#> 22               NA              NA                NA             NA
#>    field_annotation
#> 1                NA
#> 2                NA
#> 3                NA
#> 4                NA
#> 5                NA
#> 6                NA
#> 7                NA
#> 8                NA
#> 9                NA
#> 10               NA
#> 11               NA
#> 12               NA
#> 13               NA
#> 14               NA
#> 15               NA
#> 16               NA
#> 17               NA
#> 18               NA
#> 19               NA
#> 20               NA
#> 21               NA
#> 22               NA
#> 
#> $`Column names`
#>  [1] "record_id"                    "redcap_event_name"           
#>  [3] "redcap_repeat_instrument"     "redcap_repeat_instance"      
#>  [5] "cpr"                          "inclusion"                   
#>  [7] "dob"                          "age"                         
#>  [9] "age_integer"                  "sex"                         
#> [11] "cohabitation"                 "hypertension"                
#> [13] "diabetes"                     "region"                      
#> [15] "baseline_data_start_complete" "mrs_assessed"                
#> [17] "mrs_date"                     "mrs_score"                   
#> [19] "mrs_complete"                 "event_date"                  
#> [21] "event_type"                   "new_event_complete"          
#>